MongoWork Finished

+++ src/main/java/com/nearbuy/dynamic/pricing/dynamicpricing/service/model/MerchantDetail.java
@@ -29,23 +29,23 @@ public class MerchantDetail {
                     '}';
         }
 
-        public float getLatitude() {
+        public Double getLatitude() {
             return latitude;
         }
 
-        public void setLatitude(float latitude) {
+        public void setLatitude(Double latitude) {
             this.latitude = latitude;
         }
 
-        public float getLongitude() {
+        public Double getLongitude() {
             return longitude;
         }
 
-        public void setLongitude(float longitude) {
+        public void setLongitude(Double longitude) {
             this.longitude = longitude;
         }
 
-        float latitude;
-        float longitude;
+        Double latitude;
+        Double longitude;
     }
 }
